% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.seg}
\alias{example.seg}
\title{A data frame consisting of the sample ID, the location of a genomic locus, number of markers, and
an associated numeric value (logR) representing the relative copy change compared to the reference.}
\format{
An object of class \code{data.frame} with 252 rows and 6 columns.
}
\usage{
example.seg
}
\value{
segmented data
}
\description{
A data frame consisting of the sample ID, the location of a genomic locus, number of markers, and
an associated numeric value (logR) representing the relative copy change compared to the reference.
}
\keyword{datasets}
