% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_chr_bins}
\alias{hg38_chr_bins}
\title{A vector that stores the number of 1MB-size genomic bins for each chromosome from the hg38 genome.
It is used to normalize the segment lengths to the corresponding chromosomes.}
\format{
An object of class \code{numeric} of length 24.
}
\usage{
hg38_chr_bins
}
\value{
the numebr of genomic bins in the hg38 genome
}
\description{
A vector that stores the number of 1MB-size genomic bins for each chromosome from the hg38 genome.
It is used to normalize the segment lengths to the corresponding chromosomes.
}
\keyword{datasets}
