% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelseg.R
\name{labelseg}
\alias{labelseg}
\title{Annotate segment profiles}
\usage{
labelseg(data, genome = "hg38")
}
\arguments{
\item{data}{CNA segment data with logR values}

\item{genome}{genome version of input data}
}
\value{
segment data with labels
}
\description{
This function assigns segments SCNA labels to call different levels of SCNA.
It estimates calling thresholds from logR distribution of individual samples.
}
